修复 sku 列表类型问题, 添加商品编辑接口

+++ src/main/java/io/lesible/model/response/sku/SkuListInfo.java
@@ -48,13 +48,13 @@ public class SkuListInfo {
      * 二级规格 id,缺省为 0
      */
     @JsonProperty("spec_detail_id2")
-    private Integer secondSpecDetailId;
+    private Long secondSpecDetailId;
 
     /**
      * 三级规格 id,缺省为 0
      */
     @JsonProperty("spec_detail_id3")
-    private Integer thirdSpecDetailId;
+    private Long thirdSpecDetailId;
 
     /**
      * 一级规格名称
